add: `connect` and `stopListener` to `connect.ts`
update: `login`, `abort`, `checkAndReopen` and`getWebsocketInfo` methods
chore: dehydrate methods from `rocketchat.js` to `connect.ts`
update: IRocketChat interface
remove: methods from `sdk.ts`
