+++ app/lib/rocketchat/rocketchat.js
@@ -72,7 +72,11 @@ import {
 	determineAuthType,
 	disconnect,
 	checkAndReopen,
-	abort
+	abort,
+	getServerInfo,
+	getWebsocketInfo,
+	stopListener,
+	connect
 } from './services/connect';
 import * as restAPis from './services/restApi';
 
@@ -84,8 +88,7 @@ export const THEME_PREFERENCES_KEY = 'RC_THEME_PREFERENCES_KEY';
 export const CRASH_REPORT_KEY = 'RC_CRASH_REPORT_KEY';
 export const ANALYTICS_EVENTS_KEY = 'RC_ANALYTICS_EVENTS_KEY';
 export const MIN_ROCKETCHAT_VERSION = '0.70.0';
-
-const STATUSES = ['offline', 'online', 'away', 'busy'];
+export const STATUSES = ['offline', 'online', 'away', 'busy'];
 
 const RocketChat = {
 	TOKEN_KEY,
@@ -110,287 +113,14 @@ const RocketChat = {
 		}
 	},
 	canOpenRoom,
-	async getWebsocketInfo({ server }) {
-		const sdk = new RocketchatClient({ host: server, protocol: 'ddp', useSsl: useSsl(server) });
-
-		try {
-			await sdk.connect();
-		} catch (err) {
-			if (err.message && err.message.includes('400')) {
-				return {
-					success: false,
-					message: I18n.t('Websocket_disabled', { contact: I18n.t('Contact_your_server_admin') })
-				};
-			}
-		}
-
-		sdk.disconnect();
-
-		return {
-			success: true
-		};
-	},
-	async getServerInfo(server) {
-		try {
-			const response = await RNFetchBlob.fetch('GET', `${server}/api/info`, { ...RocketChatSettings.customHeaders });
-			try {
-				// Try to resolve as json
-				const jsonRes = response.json();
-				if (!jsonRes?.success) {
-					return {
-						success: false,
-						message: I18n.t('Not_RC_Server', { contact: I18n.t('Contact_your_server_admin') })
-					};
-				}
-				if (compareServerVersion(jsonRes.version, 'lowerThan', MIN_ROCKETCHAT_VERSION)) {
-					return {
-						success: false,
-						message: I18n.t('Invalid_server_version', {
-							currentVersion: jsonRes.version,
-							minVersion: MIN_ROCKETCHAT_VERSION
-						})
-					};
-				}
-				return jsonRes;
-			} catch (error) {
-				// Request is successful, but response isn't a json
-			}
-		} catch (e) {
-			if (e?.message) {
-				if (e.message === 'Aborted') {
-					reduxStore.dispatch(selectServerFailure());
-					throw e;
-				}
-				return {
-					success: false,
-					message: e.message
-				};
-			}
-		}
-
-		return {
-			success: false,
-			message: I18n.t('Not_RC_Server', { contact: I18n.t('Contact_your_server_admin') })
-		};
-	},
-	stopListener(listener) {
-		return listener && listener.stop();
-	},
+	getWebsocketInfo,
+	getServerInfo,
+	stopListener,
 	// Abort all requests and create a new AbortController
 	abort,
 	checkAndReopen,
 	disconnect,
-	connect({ server, user, logoutOnError = false }) {
-		return new Promise(resolve => {
-			if (this?.sdk?.client?.host === server) {
-				return resolve();
-			} else {
-				this.disconnect();
-				database.setActiveDB(server);
-			}
-			reduxStore.dispatch(connectRequest());
-
-			if (this.connectTimeout) {
-				clearTimeout(this.connectTimeout);
-			}
-
-			if (this.connectingListener) {
-				this.connectingListener.then(this.stopListener);
-			}
-
-			if (this.connectedListener) {
-				this.connectedListener.then(this.stopListener);
-			}
-
-			if (this.closeListener) {
-				this.closeListener.then(this.stopListener);
-			}
-
-			if (this.usersListener) {
-				this.usersListener.then(this.stopListener);
-			}
-
-			if (this.notifyAllListener) {
-				this.notifyAllListener.then(this.stopListener);
-			}
-
-			if (this.rolesListener) {
-				this.rolesListener.then(this.stopListener);
-			}
-
-			if (this.notifyLoggedListener) {
-				this.notifyLoggedListener.then(this.stopListener);
-			}
-
-			this.unsubscribeRooms();
-
-			EventEmitter.emit('INQUIRY_UNSUBSCRIBE');
-
-			this.sdk = sdk.initialize(server);
-			this.getSettings();
-
-			this.sdk
-				.connect()
-				.then(() => {
-					console.log('connected');
-				})
-				.catch(err => {
-					console.log('connect error', err);
-				});
-
-			this.connectingListener = this.sdk.onStreamData('connecting', () => {
-				reduxStore.dispatch(connectRequest());
-			});
-
-			this.connectedListener = this.sdk.onStreamData('connected', () => {
-				const { connected } = reduxStore.getState().meteor;
-				if (connected) {
-					return;
-				}
-				reduxStore.dispatch(connectSuccess());
-				const { server: currentServer } = reduxStore.getState().server;
-				if (user?.token && server === currentServer) {
-					reduxStore.dispatch(loginRequest({ resume: user.token }, logoutOnError));
-				}
-			});
-
-			this.closeListener = this.sdk.onStreamData('close', () => {
-				reduxStore.dispatch(disconnectAction());
-			});
-
-			this.usersListener = this.sdk.onStreamData(
-				'users',
-				protectedFunction(ddpMessage => RocketChat._setUser(ddpMessage))
-			);
-
-			this.notifyAllListener = this.sdk.onStreamData(
-				'stream-notify-all',
-				protectedFunction(async ddpMessage => {
-					const { eventName } = ddpMessage.fields;
-					if (/public-settings-changed/.test(eventName)) {
-						const { _id, value } = ddpMessage.fields.args[1];
-						const db = database.active;
-						const settingsCollection = db.get('settings');
-						try {
-							const settingsRecord = await settingsCollection.find(_id);
-							const { type } = defaultSettings[_id];
-							if (type) {
-								await db.action(async () => {
-									await settingsRecord.update(u => {
-										u[type] = value;
-									});
-								});
-							}
-							reduxStore.dispatch(updateSettings(_id, value));
-						} catch (e) {
-							log(e);
-						}
-					}
-				})
-			);
-
-			this.rolesListener = this.sdk.onStreamData(
-				'stream-roles',
-				protectedFunction(ddpMessage => onRolesChanged(ddpMessage))
-			);
-
-			// RC 4.1
-			this.sdk.onStreamData('stream-user-presence', ddpMessage => {
-				const userStatus = ddpMessage.fields.args[0];
-				const { uid } = ddpMessage.fields;
-				const [, status, statusText] = userStatus;
-				const newStatus = { status: STATUSES[status], statusText };
-				reduxStore.dispatch(setActiveUsers({ [uid]: newStatus }));
-
-				const { user: loggedUser } = reduxStore.getState().login;
-				if (loggedUser && loggedUser.id === uid) {
-					reduxStore.dispatch(setUser(newStatus));
-				}
-			});
-
-			this.notifyLoggedListener = this.sdk.onStreamData(
-				'stream-notify-logged',
-				protectedFunction(async ddpMessage => {
-					const { eventName } = ddpMessage.fields;
-
-					// `user-status` event is deprecated after RC 4.1 in favor of `stream-user-presence/${uid}`
-					if (/user-status/.test(eventName)) {
-						this.activeUsers = this.activeUsers || {};
-						if (!this._setUserTimer) {
-							this._setUserTimer = setTimeout(() => {
-								const activeUsersBatch = this.activeUsers;
-								InteractionManager.runAfterInteractions(() => {
-									reduxStore.dispatch(setActiveUsers(activeUsersBatch));
-								});
-								this._setUserTimer = null;
-								return (this.activeUsers = {});
-							}, 10000);
-						}
-						const userStatus = ddpMessage.fields.args[0];
-						const [id, , status, statusText] = userStatus;
-						this.activeUsers[id] = { status: STATUSES[status], statusText };
-
-						const { user: loggedUser } = reduxStore.getState().login;
-						if (loggedUser && loggedUser.id === id) {
-							reduxStore.dispatch(setUser({ status: STATUSES[status], statusText }));
-						}
-					} else if (/updateAvatar/.test(eventName)) {
-						const { username, etag } = ddpMessage.fields.args[0];
-						const db = database.active;
-						const userCollection = db.get('users');
-						try {
-							const [userRecord] = await userCollection.query(Q.where('username', Q.eq(username))).fetch();
-							await db.action(async () => {
-								await userRecord.update(u => {
-									u.avatarETag = etag;
-								});
-							});
-						} catch {
-							// We can't create a new record since we don't receive the user._id
-						}
-					} else if (/permissions-changed/.test(eventName)) {
-						const { _id, roles } = ddpMessage.fields.args[1];
-						const db = database.active;
-						const permissionsCollection = db.get('permissions');
-						try {
-							const permissionsRecord = await permissionsCollection.find(_id);
-							await db.action(async () => {
-								await permissionsRecord.update(u => {
-									u.roles = roles;
-								});
-							});
-							reduxStore.dispatch(updatePermission(_id, roles));
-						} catch (err) {
-							//
-						}
-					} else if (/Users:NameChanged/.test(eventName)) {
-						const userNameChanged = ddpMessage.fields.args[0];
-						const db = database.active;
-						const userCollection = db.get('users');
-						try {
-							const userRecord = await userCollection.find(userNameChanged._id);
-							await db.action(async () => {
-								await userRecord.update(u => {
-									Object.assign(u, userNameChanged);
-								});
-							});
-						} catch {
-							// User not found
-							await db.action(async () => {
-								await userCollection.create(u => {
-									u._raw = sanitizedRaw({ id: userNameChanged._id }, userCollection.schema);
-									Object.assign(u, userNameChanged);
-								});
-							});
-						}
-					}
-				})
-			);
-
-			resolve();
-		});
-	},
-
+	connect,
 	async shareExtensionInit(server) {
 		database.setShareDB(server);
